:boom: feat: staff and finance! :boom:
- Staffs can be added and  removed
- Made some minor enhancements in student fee receiving process
- Salary can be paid to teacher and staff

Note: Our program design is not good enough right now. We need to improve on it.
In upcoming PRs I'll try to shape the program design to be more robust and
will replace some common logical functions with a single function.
For e.g. removal of student, teacher or  staff should use single function
by passing objects and inheritance.

Closes #3
Closes #10